+++ core/src/com/unciv/models/ruleset/tile/TileImprovement.kt
@@ -27,9 +27,10 @@ class TileImprovement : NamedStats() {
 
     fun getTurnsToBuild(civInfo: CivilizationInfo): Int {
         var realTurnsToBuild = turnsToBuild.toFloat() * civInfo.gameInfo.gameParameters.gameSpeed.modifier
-        if (civInfo.containsBuildingUnique("Worker construction increased 25%"))
+        // todo UNIFY THESE
+        if (civInfo.hasUnique("Worker construction increased 25%"))
             realTurnsToBuild *= 0.75f
-        if (civInfo.policies.isAdopted("Citizenship"))
+        if (civInfo.hasUnique("Tile improvement speed +25%"))
             realTurnsToBuild *= 0.75f
         return realTurnsToBuild.roundToInt()
     }